async flow

+++ waldiez/running/running.py
@@ -158,7 +158,7 @@ async def a_install_requirements(
     printer(f"Installing requirements: {requirements_string}")
     pip_install = [sys.executable, "-m", "pip", "install"]
     if not in_virtualenv():
-        pip_install.append("--user")
+        pip_install.extend(["--user", "--break-system-packages"])
     pip_install.extend(extra_requirements)
     proc = await asyncio.create_subprocess_exec(
         *pip_install,
@@ -205,7 +205,8 @@ def after_run(
             printer("Generating mermaid sequence diagram...")
             mmd_path = temp_dir / f"{flow_name}.mmd"
             generate_sequence_diagram(events_csv_path, mmd_path)
-            shutil.copyfile(mmd_path, output_dir / f"{flow_name}.mmd")
+            if mmd_path.exists():
+                shutil.copyfile(mmd_path, output_dir / f"{flow_name}.mmd")
     if output_path:
         destination_dir = output_path.parent
         destination_dir = (
@@ -216,22 +217,59 @@ def after_run(
         destination_dir.mkdir(parents=True, exist_ok=True)
         # copy the contents of the temp dir to the destination dir
         printer(f"Copying the results to {destination_dir}")
-        temp_dir.mkdir(parents=True, exist_ok=True)
-        for item in temp_dir.iterdir():
-            # skip cache files
-            if (
-                item.name.startswith("__pycache__")
-                or item.name.endswith(".pyc")
-                or item.name.endswith(".pyo")
-                or item.name.endswith(".pyd")
-                or item.name == ".cache"
-            ):
-                continue
-            if item.is_file():
-                shutil.copy(item, destination_dir)
-            else:
-                shutil.copytree(item, destination_dir / item.name)
+        copy_results(
+            temp_dir=temp_dir,
+            output_path=output_path,
+            output_dir=output_dir,
+            destination_dir=destination_dir,
+        )
     shutil.rmtree(temp_dir)
+
+
+def copy_results(
+    temp_dir: Path,
+    output_path: Path,
+    output_dir: Path,
+    destination_dir: Path,
+) -> None:
+    """Copy the results to the output directory.
+
+    Parameters
+    ----------
+    temp_dir : Path
+        The temporary directory.
+    output_path : Path
+        The output path.
+    output_dir : Path
+        The output directory.
+    destination_dir : Path
+        The destination directory.
+    """
+    temp_dir.mkdir(parents=True, exist_ok=True)
+    for item in temp_dir.iterdir():
+        # skip cache files
+        if (
+            item.name.startswith("__pycache__")
+            or item.name.endswith(".pyc")
+            or item.name.endswith(".pyo")
+            or item.name.endswith(".pyd")
+            or item.name == ".cache"
+        ):
+            continue
+        if item.is_file():
+            shutil.copy(item, destination_dir)
+        else:
+            shutil.copytree(item, destination_dir / item.name)
+    if output_path.is_file():
+        if output_path.suffix == ".waldiez":
+            output_path = output_path.with_suffix(".py")
+        if output_path.suffix == ".py":
+            src = temp_dir / output_path.name
+            if src.exists():
+                dst = destination_dir / output_path.name
+                if dst.exists():
+                    dst.unlink()
+                shutil.copyfile(src, output_dir / output_path.name)
 
 
 def get_printer() -> Callable[..., None]:
